+++ src/scripts/shared/components/component.ts
@@ -28,7 +28,7 @@ export class Component {
     copyComponentData(obj: this): SaveObject {
         const saveObject: SaveObject = {}
         for (const key in obj) {
-            if (key !== 'body' && key !== 'cleared' && key !== 'collider' && key !== 'colliderGraphics' && key !== 'owner'  && key !== 'sprite' && key !== 'type' && key !== 'world') saveObject[key] = obj[key]
+            if (key !== 'body' && key !== 'cleared' && key !== 'collider' && key !== 'colliderGraphics' && key !== 'owner' && key !== 'sprite' && key !== 'type' && key !== 'world') saveObject[key] = obj[key]
         }
         return saveObject
     }
