+++ src/scripts/shared/entities/create-entity.ts
