+++ src/scripts/shared/entities/entity.ts
