+++ src/scripts/shared/systems/render.ts
