+++ src/scripts/shared/systems/system.ts
